[New] Added test cases for record form form subforms subform fields components violation-title
