+++ app/javascript/components/record-form/form/subforms/subform-fields/components/violation-title/component.jsx
@@ -22,8 +22,8 @@ const Component = ({ title, values, fields }) => {
   ) : null;
 
   return (
-    <div className={css.container}>
-      <div className={css.titleViolation}>
+    <div className={css.container} data-testid="violation-title">
+      <div className={css.title}>
         {title} {renderShortUniqueId}
       </div>
       <div>{renderChipStatus}</div>
